prototype_history/output_prototype_AZ/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
Pillow==10.3.0
pydantic==2.7.1
email-validator==2.1.1